# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CustomTime:
    properties:
      time.Time:
        type: string
    type: object
  models.EntityType:
    enum:
    - account
    - user
    type: string
    x-enum-varnames:
    - EntityTypeAccount
    - EntityTypeUser
...
  models.JSONMap:
    additionalProperties: true
    type: object
...
  models.JWTTask:
    properties:
      cluster_ids:
        allOf:
        - $ref: '#/definitions/models.JSONMap'
        description: Multi-cluster support
      completed_at:
        $ref: '#/definitions/models.CustomTime'
      created_at:
        $ref: '#/definitions/models.CustomTime'
      entity_id:
        type: string
      entity_type:
        $ref: '#/definitions/models.EntityType'
      error:
        type: string
      id:
        type: string
      max_retries:
        type: integer
      operation:
        $ref: '#/definitions/models.Operation'
      public_key:
        description: 冗余存储账户公钥，用于快照和追溯
        type: string
      retries:
        type: integer
      status:
        $ref: '#/definitions/models.TaskStatus'
      sync_results:
        allOf:
        - $ref: '#/definitions/models.JSONMap'
        description: 各集群同步结果
      trigger_type:
        allOf:
        - $ref: '#/definitions/models.TriggerType'
        description: 触发方式：自动或手动
      updated_at:
        $ref: '#/definitions/models.CustomTime'
    type: object
...
  models.Operation:
    enum:
    - create
    - update
    - disable
    - enable
    type: string
    x-enum-varnames:
    - OperationCreate
    - OperationUpdate
    - OperationDisable
    - OperationEnable
...
  models.TaskStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    - retrying
    type: string
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusProcessing
    - TaskStatusCompleted
    - TaskStatusFailed
    - TaskStatusRetrying
...
  models.TriggerType:
    enum:
    - auto
    - manual
    type: string
    x-enum-comments:
      TriggerTypeAuto: 自动触发（账户/用户创建/更新时）
      TriggerTypeManual: 手动触发（用户点击同步按钮）
    x-enum-descriptions:
    - 自动触发（账户/用户创建/更新时）
    - 手动触发（用户点击同步按钮）
    x-enum-varnames:
    - TriggerTypeAuto
    - TriggerTypeManual
...
  service.JWTTaskStats:
    properties:
      completed:
        type: integer
      failed:
        type: integer
      pending:
        type: integer
      processing:
        type: integer
      retrying:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing NATS control plane with multi-cluster support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NATS Control API
  version: "1.0"
paths:
...
  /jwt-tasks:
    get:
      description: Get a list of JWT tasks with optional status filter and pagination
      parameters:
      - description: Task status filter (pending, processing, completed, failed, retrying)
        in: query
        name: status
        type: string
      - description: Number of tasks to return (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Number of tasks to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/models.JWTTask'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List JWT tasks
      tags:
      - jwt-tasks
  /jwt-tasks/{id}:
    get:
      description: Retrieve a JWT task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task details
          schema:
            $ref: '#/definitions/models.JWTTask'
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get JWT task by ID
      tags:
      - jwt-tasks
  /jwt-tasks/{id}/retry:
    post:
      description: Retry a failed JWT task by resetting it to pending status
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task retry initiated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retry failed JWT task
      tags:
      - jwt-tasks
  /jwt-tasks/failed:
    get:
      description: Get a list of failed JWT tasks
      parameters:
      - description: Number of tasks to return (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of failed tasks
          schema:
            items:
              $ref: '#/definitions/models.JWTTask'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List failed JWT tasks
      tags:
      - jwt-tasks
  /jwt-tasks/stats:
    get:
      description: Get statistics of JWT tasks by status
      produces:
      - application/json
      responses:
        "200":
          description: Task statistics
          schema:
            $ref: '#/definitions/service.JWTTaskStats'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get JWT task statistics
      tags:
      - jwt-tasks
...
schemes:
- http
- https
swagger: "2.0"